+++ app/assets/stylesheets/components/_banner.scss
@@ -9,17 +9,18 @@
 }
 
 h2 {
-  font-size: 100px;
+  font-size: 70px;
   font-weight: 200;
   margin: 0.4em 0;
   color: white;
   line-height: 2.2;
-  text-align: center;
+  padding-left: 30%;
 }
 
 
 .search-form-control {
   position: relative;
+
 }
 
 .search-form-control .btn {
@@ -27,12 +28,15 @@ h2 {
   top: 8px;
   bottom: 8px;
   right: 8px;
+  border-radius: 20px;
 }
 
 .search-form-control .form-control {
   height: 3.5rem;
   box-shadow: 0 2px 6px rgba(0,0,0,0.08);
   border: 1px solid #E7E7E7;
+  border-radius: 5px;
+
 }
 
 .search-form-control .form-control:focus {
