+++ app/views/shared/_banner.html.erb
@@ -3,7 +3,7 @@
     <h2>
       <span
            class="txt-rotate"
-           data-period="2000"
+           data-period="500"
            data-rotate='[ "Need to go?", "Worry no more." ]'></span>
     </h2>
   </div>
