+++ app/views/shared/_navbar.html.erb
@@ -1,5 +1,5 @@
 <div class="navbar navbar-expand-sm navbar-light navbar-lewagon">
-  <%= link_to "#", class: "navbar-brand" do %>
+  <%= link_to"/", class: "navbar-brand" do %>
     <span class="nav-bar-text">
       <i class="fas fa-toilet fa-2x ml-1" id="icon"></i>
       Loo4U </span>
